Allow user to classify regions in the mosaic.
Running geosearch.py with "--categories c.xml" allows the user
to classify regions identified and displayed in the mosaic window, with
a keyboard interface using shortcut keys defined in c.xml. The results
are saved in "regions.xml".

The structure of c.xml should have a root node with children
 <category shortcut="q">Category Name</category>
where q is the key to be pressed to assign the currently-highlighted
thumbnail to "Category Name".